+++ src/test/java/com/getindata/kafka/connect/iceberg/sink/TestIcebergUtil.java
@@ -27,8 +27,11 @@ import org.junit.jupiter.api.io.TempDir;
 
 import java.io.IOException;
 import java.nio.file.Path;
+import java.time.*;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.ArrayList;
 import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -43,40 +46,41 @@ class TestIcebergUtil {
     final String unwrapWithGeomSchema = Testing.Files.readResourceAsString("json/serde-with-schema_geom.json");
     final String unwrapWithArraySchema = Testing.Files.readResourceAsString("json/serde-with-array.json");
     final String unwrapWithArraySchema2 = Testing.Files.readResourceAsString("json/serde-with-array2.json");
-
+    final String debeziumTimeCoercionSchema = Testing.Files.readResourceAsString("json/debezium-annotated-schema.json");
+    final String debeziumMetadataSchema = Testing.Files.readResourceAsString("json/debezium-metadata-schema.json");
     final String customPartitionColumn = Testing.Files.readResourceAsString("json/custom-partition-column.json");
+
+    private final IcebergSinkConfiguration defaultConfiguration = new IcebergSinkConfiguration(new HashMap());
 
     @Test
     public void testNestedJsonRecord() throws JsonProcessingException {
         IcebergChangeEvent e = new IcebergChangeEvent("test",
                 MAPPER.readTree(serdeWithSchema).get("payload"), null,
-                MAPPER.readTree(serdeWithSchema).get("schema"), null);
+                MAPPER.readTree(serdeWithSchema).get("schema"), null, this.defaultConfiguration);
         Schema schema = e.icebergSchema(defaultPartitionColumn);
-        assertTrue(schema.toString().contains("before: optional struct<2: id: optional int, 3: first_name: optional string, " +
-                "4:"));
+        assertTrue(schema.toString().contains("before: optional struct<2: id: optional int (), " +
+                "3: first_name: optional string (), 4:"));
     }
 
     @Test
     public void testUnwrapJsonRecord() throws IOException {
         IcebergChangeEvent e = new IcebergChangeEvent("test",
                 MAPPER.readTree(unwrapWithSchema).get("payload"), null,
-                MAPPER.readTree(unwrapWithSchema).get("schema"), null);
+                MAPPER.readTree(unwrapWithSchema).get("schema"), null, this.defaultConfiguration);
         Schema schema = e.icebergSchema(defaultPartitionColumn);
         GenericRecord record = e.asIcebergRecord(schema, defaultPartitionColumn, defaultPartitionTimestamp);
         assertEquals("orders", record.getField("__table").toString());
         assertEquals(16850, record.getField("order_date"));
-        System.out.println(schema);
-        System.out.println(record);
     }
 
     @Test
     public void testNestedArrayJsonRecord() throws JsonProcessingException {
         IcebergChangeEvent e = new IcebergChangeEvent("test",
                 MAPPER.readTree(unwrapWithArraySchema).get("payload"), null,
-                MAPPER.readTree(unwrapWithArraySchema).get("schema"), null);
+                MAPPER.readTree(unwrapWithArraySchema).get("schema"), null, this.defaultConfiguration);
         Schema schema = e.icebergSchema(defaultPartitionColumn);
-        assertTrue(schema.asStruct().toString().contains("struct<1: name: optional string, 2: pay_by_quarter: optional list<int>, 4: schedule: optional list<string>, 6:"));
-        System.out.println(schema.asStruct());
+        assertTrue(schema.asStruct().toString().contains("struct<1: name: optional string (), " +
+                "2: pay_by_quarter: optional list<int> (), 4: schedule: optional list<string> (), 6:"));
         System.out.println(schema.findField("pay_by_quarter").type().asListType().elementType());
         System.out.println(schema.findField("schedule").type().asListType().elementType());
         assertEquals(schema.findField("pay_by_quarter").type().asListType().elementType().toString(), "int");
@@ -90,7 +94,7 @@ class TestIcebergUtil {
         assertThrows(RuntimeException.class, () -> {
             IcebergChangeEvent e = new IcebergChangeEvent("test",
                     MAPPER.readTree(unwrapWithArraySchema2).get("payload"), null,
-                    MAPPER.readTree(unwrapWithArraySchema2).get("schema"), null);
+                    MAPPER.readTree(unwrapWithArraySchema2).get("schema"), null, this.defaultConfiguration);
             Schema schema = e.icebergSchema(defaultPartitionColumn);
             System.out.println(schema.asStruct());
             System.out.println(schema);
@@ -103,10 +107,11 @@ class TestIcebergUtil {
     public void testNestedGeomJsonRecord() throws JsonProcessingException {
         IcebergChangeEvent e = new IcebergChangeEvent("test",
                 MAPPER.readTree(unwrapWithGeomSchema).get("payload"), null,
-                MAPPER.readTree(unwrapWithGeomSchema).get("schema"), null);
+                MAPPER.readTree(unwrapWithGeomSchema).get("schema"), null, this.defaultConfiguration);
         Schema schema = e.icebergSchema(defaultPartitionColumn);
+
         GenericRecord record = e.asIcebergRecord(schema, defaultPartitionColumn, defaultPartitionTimestamp);
-        assertTrue(schema.toString().contains("g: optional struct<3: wkb: optional string, 4: srid: optional int>"));
+        assertTrue(schema.toString().contains("g: optional struct<3: wkb: optional string (), 4: srid: optional int ()>"));
         GenericRecord g = (GenericRecord) record.getField("g");
         GenericRecord h = (GenericRecord) record.getField("h");
         assertEquals("AQEAAAAAAAAAAADwPwAAAAAAAPA/", g.get(0, Types.StringType.get().typeId().javaClass()));
@@ -119,7 +124,7 @@ class TestIcebergUtil {
     public void testConvertPartitionTimestampRecord() throws IOException {
         IcebergChangeEvent e = new IcebergChangeEvent("test",
                 MAPPER.readTree(customPartitionColumn).get("payload"), null,
-                MAPPER.readTree(customPartitionColumn).get("schema"), null);
+                MAPPER.readTree(customPartitionColumn).get("schema"), null, this.defaultConfiguration);
         Schema schema = e.icebergSchema(defaultPartitionColumn);
         GenericRecord record = e.asIcebergRecord(schema, defaultPartitionColumn, "timestamp");
         assertEquals("2023-03-20T18:25:27.865Z", record.getField(defaultPartitionColumn).toString());
@@ -132,7 +137,7 @@ class TestIcebergUtil {
     public void testConvertPartitionColumnRecord() throws IOException {
         IcebergChangeEvent e = new IcebergChangeEvent("test",
                 MAPPER.readTree(customPartitionColumn).get("payload"), null,
-                MAPPER.readTree(customPartitionColumn).get("schema"), null);
+                MAPPER.readTree(customPartitionColumn).get("schema"), null, this.defaultConfiguration);
         Schema schema = e.icebergSchema("timestamp");
         GenericRecord record = e.asIcebergRecord(schema, "timestamp", "timestamp");
         assertEquals("2023-03-20T18:25:27.865Z", record.getField("timestamp").toString());
@@ -160,6 +165,123 @@ class TestIcebergUtil {
         JsonNode deserializedSchema = valueSerde.deserializer().deserialize("xx", serdeWithSchema.getBytes());
         System.out.println(deserializedSchema);
         assertFalse(deserializedSchema.has("schema"));
+    }
+
+    private void assertPrimitiveTemporalValues(IcebergChangeEvent event) {
+        Schema schema = event.icebergSchema(defaultPartitionColumn);
+
+        Types.NestedField ship_date = schema.findField("ship_date");
+        assertEquals(Types.IntegerType.get(), ship_date.type());
+        assertEquals("io.debezium.time.Date", ship_date.doc());
+
+        Types.NestedField ship_timestamp = schema.findField("ship_timestamp");
+        assertEquals(Types.LongType.get(), ship_timestamp.type());
+        assertEquals("io.debezium.time.MicroTimestamp", ship_timestamp.doc());
+
+        Types.NestedField ship_timestamp_zoned = schema.findField("ship_timestamp_zoned");
+        assertEquals(Types.StringType.get(), ship_timestamp_zoned.type());
+        assertEquals("io.debezium.time.ZonedTimestamp", ship_timestamp_zoned.doc());
+
+        Types.NestedField ship_time = schema.findField("ship_time");
+        assertEquals(Types.LongType.get(), ship_time.type());
+        assertEquals("io.debezium.time.MicroTime", ship_time.doc());
+
+        Types.NestedField ship_time_zoned = schema.findField("ship_time_zoned");
+        assertEquals(Types.StringType.get(), ship_time_zoned.type());
+        assertEquals("io.debezium.time.ZonedTime", ship_time_zoned.doc());
+
+        GenericRecord record = event.asIcebergRecord(schema, defaultPartitionColumn, defaultPartitionTimestamp);
+        assertEquals(record.getField("ship_date"), 77663);
+        assertEquals(record.getField("ship_timestamp"), 6710075456016196L);
+        assertEquals(record.getField("ship_timestamp_zoned"), "2023-04-11T20:32:46.821144Z");
+        assertEquals(record.getField("ship_time"), 73966821144L);
+        assertEquals(record.getField("ship_time_zoned"), "20:32:46.821144Z");
+    }
+
+    @Test
+    public void listStructSchemaHandling()
+      throws JsonProcessingException {
+        IcebergChangeEvent e = new IcebergChangeEvent("test",
+                MAPPER.readTree(debeziumMetadataSchema).get("payload"), null,
+                MAPPER.readTree(debeziumMetadataSchema).get("schema"), null,
+                defaultConfiguration
+        );
+        Schema schema = e.icebergSchema(defaultPartitionColumn);
+        String schemaString = schema.toString();
+
+        GenericRecord record = e.asIcebergRecord(schema, defaultPartitionColumn, defaultPartitionTimestamp);
+
+        assertTrue(schemaString.contains("data_collections: optional list<struct"));
+
+        GenericRecord innerRecord = (GenericRecord) ((ArrayList) record.getField("data_collections")).get(0);
+        Object value = innerRecord.getField("data_collection");
+        assertEquals("public.mine", value);
+
+        value = innerRecord.getField("event_count");
+        assertEquals(1L, value);
+
+        value = record.getField("status");
+        assertEquals("END", value);
+
+        value = record.getField("id");
+        assertEquals("12117:67299632", value);
+
+        value = record.getField("ts_ms");
+        assertEquals(1680821545908L, value);
+    }
+
+    @Test
+    public void coerceDebeziumTemporalTypesDefaultBehavior()
+      throws JsonProcessingException {
+        IcebergChangeEvent event = new IcebergChangeEvent(
+                "test",
+                MAPPER.readTree(debeziumTimeCoercionSchema).get("payload"), null,
+                MAPPER.readTree(debeziumTimeCoercionSchema).get("schema"), null,
+                this.defaultConfiguration
+        );
+
+        assertPrimitiveTemporalValues(event);
+    }
+
+    @Test
+    public void coerceDebeziumTemporalTypesDisabledBehavior(@TempDir Path localWarehouseDir)
+      throws JsonProcessingException {
+        IcebergSinkConfiguration config = TestConfig.builder()
+                .withLocalCatalog(localWarehouseDir)
+                .withCustomProperty("rich-temporal-types", "false")
+                .build();
+        IcebergChangeEvent event = new IcebergChangeEvent(
+                "test",
+                MAPPER.readTree(debeziumTimeCoercionSchema).get("payload"), null,
+                MAPPER.readTree(debeziumTimeCoercionSchema).get("schema"), null,
+                config
+        );
+
+        assertPrimitiveTemporalValues(event);
+    }
+
+    @Test
+    public void coerceDebeziumTemporalTypesEnabledBehavior(@TempDir Path localWarehouseDir)
+      throws JsonProcessingException {
+        IcebergSinkConfiguration configuration = TestConfig.builder()
+                .withLocalCatalog(localWarehouseDir)
+                .withCustomProperty("rich-temporal-types", "true")
+                .build();
+        IcebergChangeEvent e = new IcebergChangeEvent(
+                "test",
+                MAPPER.readTree(debeziumTimeCoercionSchema).get("payload"), null,
+                MAPPER.readTree(debeziumTimeCoercionSchema).get("schema"), null,
+                configuration
+        );
+        Schema schema = e.icebergSchema(defaultPartitionColumn);
+        GenericRecord record = e.asIcebergRecord(schema, defaultPartitionColumn, defaultPartitionTimestamp);
+        String schemaString = schema.toString();
+        String recordString = record.toString();
+
+        assertTrue(schemaString.contains("ship_date: optional date (io.debezium.time.Date)"));
+        assertTrue(schemaString.contains("ship_timestamp: optional timestamp (io.debezium.time.MicroTimestamp)"));
+        assertTrue(recordString.contains("2182-08-20"));
+        assertTrue(recordString.contains("2182-08-19T21:50:56.016196"));
     }
 
     @Test