+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
+-   Updates AWS lakeformation transitive dependency providing lakeformation support in s3 iceberg tables.
+
 ## [Unreleased]
+
+## [0.4.0] - 2023-05-25
+
+-   Updates dependencies to resolve some jackson-databind critical CVEs.
+-   Updates AWS lakeformation transitive dependency providing lakeformation support in s3 iceberg tables.
+-   Added Iceberg coercion support for Avro Array<Struct> types. Supports Debezium `data_collections` metadata.
+-   Added support for coercion of five Debezium temporal types to their Iceberg equivalents: Date, MicroTimestamp, ZonedTimestamp, MicroTime, and ZonedTime
+-   Rich temporal types are toggled on by new boolean configuration property: `rich-temporal-types`
 
 ## [0.3.1] - 2023-04-06
 
@@ -51,7 +61,9 @@ Kafka Connect >= 3.2.3 has updated the jackson version to an incompatible minor 
 
 -   First release
 
-[Unreleased]: https://github.com/getindata/kafka-connect-iceberg-sink/compare/0.3.1...HEAD
+[Unreleased]: https://github.com/getindata/kafka-connect-iceberg-sink/compare/0.4.0...HEAD
+
+[0.4.0]: https://github.com/getindata/kafka-connect-iceberg-sink/compare/0.3.1...0.4.0
 
 [0.3.1]: https://github.com/getindata/kafka-connect-iceberg-sink/compare/0.3.0...0.3.1